cleanup: we don't care if there is nothing to cleanup anymore.

+++ js/client/modules/@arangodb/testsuites/dump.js
@@ -305,7 +305,7 @@ class DumpRestoreHelper extends tu.runInArangoshRunner {
   }
 
   extractResults() {
-    if (this.fn !== undefined) {
+    if (this.fn !== undefined && fs.exists(this.fn)) {
       fs.remove(this.fn);
     }
     if (this.instanceManager === null) {
